Change the layout and order of code
In order to reduce the size of the script and also for better readabily and future maintenance, remove all blank lines and also line breaks in playSingleRound().

In addition, In order to make the click button's event listener look better, each line of code that is related to gameReport() or scoreReactor() should be moved into it's related scopre.