Add dynamic colored game Commentary
Make each line of game commentary i.e. round reports, get colored based on round's result. If user if lost, text color will be red. If won, color is green and for draw there is default color.

This is done by selecting the last element added to the report box. Then based on a keyword found in the line (using arr.includ() and specific keywords of "You" and "Computer" in each line) the color is determined.