Style reset button and modify buttons section layout
In order to make reset button's layout to be similar to other buttons, Add their styles to it.

Also to have a more organized layout and put reset button at below and to far right corner of three main buttons, put all sections of grey area in main into a div so style them as a flex container. This way, more controll is gained on reset button's position.