Fix reset button acting as a player choice button
Reset button is acting as one of the three selection buttons, Meaning it would initiate the game but obviously that not intended.

The reason is that the DOM element selectors which contributes in game initiation, uses "button" as selector which would select all buttons in HTML.

To fix this issue, the selector for DOM manipulations was changed to a new edited class of player selection buttons. also in order to prevent further bugs like this, the button's container class and reset button's class were also replaced by a more distinct name.