Make reset button appear only when game is started
It doesn't make sense for reset button to get shown on the page at the stage when game is not initiated. It more makes sense that a user would want to reset the game only when a few rounds are played.

In order to make reset button appear only after first click of the player choices, add a "display: none" style to it. Then by manipulating DOM make it appear (display: initial) only after one of player choice buttons is clicked.