Change color and position of reset button
The reset button seems a bit off relative to buttons and dimensions in main section. Reduce padding of buttons container in order to decrease the vertical distance between 3 game buttons and the reset button.

Also put the reset button inside a div and define a width for that div so reset button gets closer to the center of page. Additionaly, define bottom padding for the button to make that section look better.

In the end, change coloring styles of reset button to apply more distiction between this element and 3 game buttons.