Add styles to replay button in end game
The replay button that shows up in end game state doens't have any styles. This button is going to get the same style as game buttons.

First define and seperate a new class in HTML which is only used for styling buttons. Then assing this class to the replay button generated at the end game.